# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Kids Art Tales API
  description: API for children's drawing app with image-to-story, image-to-cartoon, and image-to-animation features
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /api/community/works:
    get:
      summary: List community works
      tags:
        - Community
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [audio, picture, animation]
        - in: query
          name: authorId
          schema:
            type: string
      responses:
        '200':
          description: List of works
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Work'
    post:
      summary: Publish work to community
      tags:
        - Community
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - authorId
                - type
                - mediaUrl
              properties:
                authorId:
                  type: string
                authorName:
                  type: string
                authorAvatar:
                  type: string
                type:
                  type: string
                  enum: [audio, picture, animation]
                title:
                  type: string
                description:
                  type: string
                mediaUrl:
                  type: string
                thumbnailUrl:
                  type: string
                canCoCreate:
                  type: boolean
                  default: true
      responses:
        '201':
          description: Work published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Work'
  /api/community/works/{id}:
    get:
      summary: Get work by ID
      tags:
        - Community
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Work details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Work'
        '404':
          description: Work not found
  /api/community/works/{id}/like:
    post:
      summary: Like a work
      tags:
        - Community
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: string
      responses:
        '200':
          description: Work liked
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  likeCount:
                    type: integer
                  authorReward:
                    type: integer
  /api/community/works/{id}/unlike:
    post:
      summary: Unlike a work
      tags:
        - Community
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: string
      responses:
        '200':
          description: Work unliked
  /api/community/works/{id}/liked:
    get:
      summary: Check if user liked a work
      tags:
        - Community
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Like status
          content:
            application/json:
              schema:
                type: object
                properties:
                  liked:
                    type: boolean
  /api/community/works/{id}/share:
    post:
      summary: Share a work
      tags:
        - Community
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Work shared
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  shareCount:
                    type: integer
                  shareUrl:
                    type: string
  /api/community/works/{id}/cocreate:
    post:
      summary: Co-create from a work
      tags:
        - Community
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - authorId
              properties:
                authorId:
                  type: string
                authorName:
                  type: string
                authorAvatar:
                  type: string
                title:
                  type: string
                description:
                  type: string
                mediaUrl:
                  type: string
                thumbnailUrl:
                  type: string
      responses:
        '201':
          description: Co-created work
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Work'
        '403':
          description: Co-create disabled for this work
  /api/community/users/{userId}/liked:
    get:
      summary: Get user's liked works
      tags:
        - Community
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Liked works
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Work'
  /api/community/users/{userId}/like-rewards:
    get:
      summary: Get author's like rewards
      tags:
        - Community
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Like rewards summary

  # Subscription APIs
...
components:
  schemas:
...
    Work:
      type: object
      properties:
        id:
          type: string
        authorId:
          type: string
        authorName:
          type: string
        authorAvatar:
          type: string
        type:
          type: string
          enum: [audio, picture, animation]
        title:
          type: string
        description:
          type: string
        likeCount:
          type: integer
        likedBy:
          type: array
          items:
            type: string
        canCoCreate:
          type: boolean
        mediaUrl:
          type: string
        thumbnailUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        shareCount:
          type: integer
...